requirements.txt
Faker==30.8.2
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.2